Work in MPSC
There was a challenge in the MPSC that a pointer
to the next buffer had to be stored. This worked in
Java because everything is object, but with .NET
that has value types things got more difficult.

I opted to try storing everyting in the buffers as
Tuples, and one side of the tuple store the next
buffer pointer. This needs to have some unit tests
written because it's probably got some bugs in it.